refactor: make tests easily extensible to other DBs
